# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/asset": {
            "get": {
                "description": "Returns a paginated list of assets available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get All Asset",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetAllAsset"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an asset",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Post Create Asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Name",
                        "name": "asset_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Description",
                        "name": "asset_desc",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Merk",
                        "name": "asset_merk",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Category",
                        "name": "asset_category",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Asset Price",
                        "name": "asset_price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Status",
                        "name": "asset_status",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Asset Image (JPG,PNG,JPEG)",
                        "name": "asset_image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseCreateAsset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/deleted": {
            "get": {
                "description": "Returns a list of deleted assets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get Deleted Asset",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetDeletedAsset"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/destroy/{id}": {
            "delete": {
                "description": "Permanentally Delete Asset By Id",
                "tags": [
                    "Asset"
                ],
                "summary": "Hard Delete Asset By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of asset",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseHardDeleteAssetById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/asset/most-context/{targetCol}": {
            "get": {
                "description": "Returns a list of most appear item in asset maintenance by given field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Get Most Context Asset Maintenance",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target Column to Analyze (such as: maintenance_day)",
                        "name": "targetCol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseGetMostContext"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/asset/recover/{id}": {
            "put": {
                "description": "Recover Deleted Asset By Id",
                "tags": [
                    "Asset"
                ],
                "summary": "Recover Put Deleted Asset By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of asset",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseRecoverDeleteAssetById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/asset/{id}": {
            "put": {
                "description": "Update an asset by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Put Update Asset By Id",
                "parameters": [
                    {
                        "description": "Update Asset Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RequestUpdateAssetById"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Id of asset",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseUpdateAssetById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Asset By Id",
                "tags": [
                    "Asset"
                ],
                "summary": "Soft Delete Asset By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of asset",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseSoftDeleteAssetById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Asset": {
            "type": "object",
            "properties": {
                "asset_category": {
                    "type": "string"
                },
                "asset_desc": {
                    "type": "string"
                },
                "asset_image_url": {
                    "type": "string"
                },
                "asset_merk": {
                    "type": "string"
                },
                "asset_name": {
                    "type": "string"
                },
                "asset_price": {
                    "type": "string"
                },
                "asset_status": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "description": "FK - Admin",
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entity.Metadata": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "entity.RequestUpdateAssetById": {
            "type": "object",
            "required": [
                "asset_category",
                "asset_desc",
                "asset_merk",
                "asset_name",
                "asset_price",
                "asset_status"
            ],
            "properties": {
                "asset_category": {
                    "type": "string"
                },
                "asset_desc": {
                    "type": "string"
                },
                "asset_merk": {
                    "type": "string"
                },
                "asset_name": {
                    "type": "string"
                },
                "asset_price": {
                    "type": "string"
                },
                "asset_status": {
                    "type": "string"
                }
            }
        },
        "entity.ResponseCreateAsset": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset created successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetAllAsset": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Asset"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset fetched"
                },
                "metadata": {
                    "$ref": "#/definitions/entity.Metadata"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetDeletedAsset": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Asset"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "deleted asset fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseGetMostContext": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.StatsContextTotal"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "asset fetched"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseHardDeleteAssetById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset permanentally deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.ResponseRecoverDeleteAssetById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset recovered"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "entity.ResponseSoftDeleteAssetById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "entity.ResponseUpdateAssetById": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "asset updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "entity.StatsContextTotal": {
            "type": "object",
            "properties": {
                "context": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}